Remove misuses from the TSE17 experiment.
Though technically these misuses are true positives identified by the detectors in the TSE experiments, we did not include them in the respective recall experiment, because we considered only one misuse of the same type per method (as opposed to the multiple occurrences that are actually there). Removed these from the dataset to keep it consistent.